back-end/requirements.txt
urllib3==1.24.3
openai==1.25.1
python-dotenv==1.0.0
langchain==0.1.17
Django==4.2.1
aiohttp==3.8.3
Brotli==1.0.9
flask==2.3.3
gunicorn==21.2.0
gevent==23.9.0.post1
edge-tts==6.1.8
djangorestframework==3.14.0
django-cors-headers==4.2.0
channels==4.0.0
daphne==4.0.0
schedule==1.2.0
drf-yasg==1.21.7
pymilvus==2.2.12
datetime==5.2
transformers==4.33.1
sentence-transformers==2.2.2
easygoogletranslate==0.0.4
volcengine==1.0.103
requests==2.31.0
jieba==0.42.1
websockets==11.0.3
APScheduler==3.10.4
faiss-cpu
langsmith==0.1.54
langchain-openai==0.1.6
litellm==1.34.6